Move Jinja2 environment init code from config file to app init file
We want to keep code out of the config file.
